+++ com.vmturbo.market.component/src/main/java/com/vmturbo/market/cloudscaling/sma/entities/SMAInput.java
@@ -598,7 +598,7 @@ public class SMAInput {
         }
         ComputeTierInfo computeTierInfo = computeTier.getTypeSpecificInfo().getComputeTier();
         String family = computeTierInfo.getFamily();
-        float coupons = computeTierInfo.getNumCoupons();
+        float coupons = (float)computeTierInfo.getNumCoupons();
 
         /*
          * Find all the contexts in which this template belongs.  Iterate threw osTypes and lookup
